feat: add more test suites to NavLink testing
